extract cocoapods plugin to separated id
