+++ src/main/resources/META-INF/gradle-plugins/dev.icerock.mobile.multiplatform.cocoapods.properties
@@ -1,0 +1,1 @@
+implementation-class=dev.icerock.gradle.CocoapodsPlugin
